Ensure that Bank holidays aren't clickable
When clicking on a bank holiday entry on the calendar page, the client
side redirect was throwing an error as it was trying to do a find on a
vacation rather than on an instance of bank holiday.

Added a check in the clientside to determine the type of calendar entry
and take the appropriate action - in this case redirect to the holiday
request or nothing if it's a bank holiday entry.