Calendar persistence through HTML5 history api
Added client side event handlers to calendar navigation which adds the
current month and year to the url when these events are fired. This adds
persistence to the calendar page for any browsers which support the HTML5
history api, so specific dates can be linked to, and immediately shown
on the page.

For example if a user goes to /calendar/show?year=2012&month=2 the
calendar they'll see is for March 2012(months start from 0 in JavaScript).